+++ tests/fixtures/source/WrongReturnMain.xs
@@ -1,0 +1,4 @@
+// SPDX-FileCopyrightText: 2026 Leitwolf <xs-lang.chess031@slmails.com>
+// SPDX-License-Identifier: Apache-2.0
+
+fn main() => Int { return 0; }