+++ xs/include/xs/mir.h
@@ -55,6 +55,7 @@ typedef enum
 typedef enum
 {
   XS_MIR_INSTRUCTION_CONST_I64,
+  XS_MIR_INSTRUCTION_CONST_I32,
   XS_MIR_INSTRUCTION_ADD_I64,
   XS_MIR_INSTRUCTION_LOAD,
   XS_MIR_INSTRUCTION_STORE,
@@ -110,6 +111,7 @@ XsMirStatus xs_mir_block_set_branch(XsMirBlock *block, XsMirValueId condition, c
                                     const XsMirBlock *else_target, XsMirError *error);
 XsMirStatus xs_mir_block_set_unreachable(XsMirBlock *block, XsMirError *error);
 XsMirStatus xs_mir_block_add_const_i64(XsMirBlock *block, int64_t value, XsMirValueId *result, XsMirError *error);
+XsMirStatus xs_mir_block_add_const_i32(XsMirBlock *block, int32_t value, XsMirValueId *result, XsMirError *error);
 XsMirStatus xs_mir_block_add_i64(XsMirBlock *block, XsMirValueId left, XsMirValueId right, XsMirValueId *result,
                                  XsMirError *error);
 XsMirStatus xs_mir_block_add_load(XsMirBlock *block, const XsMirPlace *place, XsMirType result_type,